Fix bigfiles issue
